Standardize repository layout
In general, you don't want your source files located in a `src/` directory
[1]. `go build` uses the directory name for the default output file, and
doing something like:

    cd src
    go build

Results in an executable named `src`, which isn't terribly useful as a
name. Moving everything up one level means that we can run `go build` in
the top level directory and get `prom-keycloak-proxy` as the result (and
is generally a more typical layout for a Go project).

[1]: https://github.com/golang-standards/project-layout#directories-you-shouldnt-have
