+++ go.mod
@@ -3,14 +3,15 @@ module github.com/OCP-on-NERC/prom-keycloak-proxy
 go 1.20
 
 require (
-	github.com/Nerzal/gocloak/v13 v13.9.0 // indirect
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible // indirect
+	github.com/Nerzal/gocloak/v13 v13.9.0
+	github.com/gorilla/mux v1.8.1
+)
+
+require (
 	github.com/go-resty/resty/v2 v2.7.0 // indirect
 	github.com/golang-jwt/jwt/v5 v5.0.0 // indirect
-	github.com/gorilla/mux v1.8.1 // indirect
 	github.com/opentracing/opentracing-go v1.2.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/segmentio/ksuid v1.0.4 // indirect
 	golang.org/x/net v0.17.0 // indirect
-	gopkg.in/resty.v1 v1.10.3 // indirect
 )
